+++ CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## Unreleased
 
+## 3.2.0
+
 ### Added
 
+* Close poll op frame
+* Update poll end time op frame
 * Invest fee
+* Voting tests
 
 ### Changed